+++ uniffi_bindgen/src/interface/attributes.rs
@@ -28,12 +28,13 @@ pub(super) enum Attribute {
     Error,
     Name(String),
     SelfType(SelfType),
-    Threadsafe, // N.B. the `[Threadsafe]` attribute is deprecated and will be removed
     Throws(String),
     // `[External="crate_name"]` - We can `use crate_name::...` for the type.
     External(String),
     // Custom type on the scaffolding side
     Custom,
+    // The interface described is implemented as a trait.
+    Trait,
 }
 
 impl Attribute {
@@ -58,8 +59,8 @@ impl TryFrom<&weedle::attribute::ExtendedAttribute<'_>> for Attribute {
                 "ByRef" => Ok(Attribute::ByRef),
                 "Enum" => Ok(Attribute::Enum),
                 "Error" => Ok(Attribute::Error),
-                "Threadsafe" => Ok(Attribute::Threadsafe),
                 "Custom" => Ok(Attribute::Custom),
+                "Trait" => Ok(Attribute::Trait),
                 _ => anyhow::bail!("ExtendedAttributeNoArgs not supported: {:?}", (attr.0).0),
             },
             // Matches assignment-style attributes like ["Throws=Error"]
@@ -245,10 +246,8 @@ impl InterfaceAttributes {
         self.0.iter().any(|attr| attr.is_error())
     }
 
-    pub fn threadsafe(&self) -> bool {
-        self.0
-            .iter()
-            .any(|attr| matches!(attr, Attribute::Threadsafe))
+    pub fn is_trait(&self) -> bool {
+        self.0.iter().any(|attr| matches!(attr, Attribute::Trait))
     }
 }
 
@@ -260,10 +259,10 @@ impl TryFrom<&weedle::attribute::ExtendedAttributeList<'_>> for InterfaceAttribu
         let attrs = parse_attributes(weedle_attributes, |attr| match attr {
             Attribute::Enum => Ok(()),
             Attribute::Error => Ok(()),
-            Attribute::Threadsafe => Ok(()),
+            Attribute::Trait => Ok(()),
             _ => bail!(format!("{:?} not supported for interface definition", attr)),
         })?;
-        // Can't be both `[Threadsafe]` and an `[Enum]`.
+        // Can't be both `[Trait]` and an `[Enum]`.
         if attrs.len() > 1 {
             bail!("conflicting attributes on interface definition");
         }
@@ -503,10 +502,10 @@ mod test {
     }
 
     #[test]
-    fn test_threadsafe() -> Result<()> {
-        let (_, node) = weedle::attribute::ExtendedAttribute::parse("Threadsafe").unwrap();
+    fn test_trait() -> Result<()> {
+        let (_, node) = weedle::attribute::ExtendedAttribute::parse("Trait").unwrap();
         let attr = Attribute::try_from(&node)?;
-        assert!(matches!(attr, Attribute::Threadsafe));
+        assert!(matches!(attr, Attribute::Trait));
         Ok(())
     }
 
@@ -663,14 +662,14 @@ mod test {
     }
 
     #[test]
-    fn test_threadsafe_attribute() {
-        let (_, node) = weedle::attribute::ExtendedAttributeList::parse("[Threadsafe]").unwrap();
+    fn test_trait_attribute() {
+        let (_, node) = weedle::attribute::ExtendedAttributeList::parse("[Trait]").unwrap();
         let attrs = InterfaceAttributes::try_from(&node).unwrap();
-        assert!(matches!(attrs.threadsafe(), true));
+        assert!(attrs.is_trait());
 
         let (_, node) = weedle::attribute::ExtendedAttributeList::parse("[]").unwrap();
         let attrs = InterfaceAttributes::try_from(&node).unwrap();
-        assert!(matches!(attrs.threadsafe(), false));
+        assert!(!attrs.is_trait());
     }
 
     #[test]
@@ -683,12 +682,11 @@ mod test {
         let attrs = InterfaceAttributes::try_from(&node).unwrap();
         assert!(matches!(attrs.contains_enum_attr(), false));
 
-        let (_, node) = weedle::attribute::ExtendedAttributeList::parse("[Threadsafe]").unwrap();
+        let (_, node) = weedle::attribute::ExtendedAttributeList::parse("[Trait]").unwrap();
         let attrs = InterfaceAttributes::try_from(&node).unwrap();
         assert!(matches!(attrs.contains_enum_attr(), false));
 
-        let (_, node) =
-            weedle::attribute::ExtendedAttributeList::parse("[Threadsafe, Enum]").unwrap();
+        let (_, node) = weedle::attribute::ExtendedAttributeList::parse("[Trait, Enum]").unwrap();
         let err = InterfaceAttributes::try_from(&node).unwrap_err();
         assert_eq!(
             err.to_string(),
@@ -698,8 +696,7 @@ mod test {
 
     #[test]
     fn test_other_attributes_not_supported_for_interfaces() {
-        let (_, node) =
-            weedle::attribute::ExtendedAttributeList::parse("[Threadsafe, ByRef]").unwrap();
+        let (_, node) = weedle::attribute::ExtendedAttributeList::parse("[Trait, ByRef]").unwrap();
         let err = InterfaceAttributes::try_from(&node).unwrap_err();
         assert_eq!(
             err.to_string(),