+++ CHANGELOG.md
@@ -14,6 +14,11 @@
   `Contstructor` to `throws_str()`.  Added a new `throws()` method that returns
   a boolean.
 
+- removed the long deprecated `ThreadSafe` attribute.
+
+### What's changed
+- traits [todo: fill me out! Also needs docs...]
+
 ## v0.20.0 - (_2022-09-13_)
 
 [All changes in v0.20.0](https://github.com/mozilla/uniffi-rs/compare/v0.19.6...v0.20.0).